Convert model output to tidy tibble/dataframe.

Currently this only tidies up PLS model objects. The main important output objects from the PLS model are:

tidy_up(model, output, ...)

Arguments

model
The model object.

output
Which output to choose from model.

...
Not currently used. For later method additions.

Value

Tibble object with tidied model output. There are several output options:

default: Tibble with five columns for the x variables, components, loadings, scores to variable correlations, and explained variance for each component and x variable combination.

loadings, score_corr: Tibble with three columns for x variables, components, and either loadings or score to variable correlations.

explained_var: Tibble with two columns for component and it's explained variance.

scores: Tibble with one column for each component, with values for the scores for each observation.

Details

Scores: These are the individual scores calculated from the model for each observation. Use these to look for patterns between components or between X or Y variables.

Loadings: These are the combined weights in the model (including both X and Y). Strongly correlated X variables that underlie Y will have similar loadings.

Explained variance: This is the amount of variance that an individual component explains within X. This is useful to use to see which components to keep.

See also

Examples

library(pls)

#> 
#> Attaching package: ‘pls’

#> The following object is masked from ‘package:stats’:
#> 
#>     loadings

data(yarn)

NIR <- yarn$NIR
density <- yarn$density
model <- plsr(density ~ NIR)
tidy_up(model)

#> # A tibble: 7,236 x 5
#>    xvariables components  loadings scores.to.var.corr explained.variance
#>    <chr>      <chr>          <dbl>              <dbl>              <dbl>
#>  1 NIR1       component 1  0.00662             0.246                46.8
#>  2 NIR2       component 1  0.0213              0.461                46.8
#>  3 NIR3       component 1  0.0391              0.546                46.8
#>  4 NIR4       component 1  0.0518              0.544                46.8
#>  5 NIR5       component 1  0.0538              0.480                46.8
#>  6 NIR6       component 1  0.0461              0.375                46.8
#>  7 NIR7       component 1  0.0316              0.239                46.8
#>  8 NIR8       component 1  0.0115              0.0818               46.8
#>  9 NIR9       component 1 -0.0117             -0.0776               46.8
#> 10 NIR10      component 1 -0.0354             -0.213                46.8
#> # … with 7,226 more rows

tidy_up(model, "loadings")

#> # A tibble: 7,236 x 3
#>    xvariables components  loadings
#>    <chr>      <chr>          <dbl>
#>  1 NIR1       component 1  0.00662
#>  2 NIR2       component 1  0.0213 
#>  3 NIR3       component 1  0.0391 
#>  4 NIR4       component 1  0.0518 
#>  5 NIR5       component 1  0.0538 
#>  6 NIR6       component 1  0.0461 
#>  7 NIR7       component 1  0.0316 
#>  8 NIR8       component 1  0.0115 
#>  9 NIR9       component 1 -0.0117 
#> 10 NIR10      component 1 -0.0354 
#> # … with 7,226 more rows

tidy_up(model, "scores")

#> # A tibble: 28 x 27
#>    scores.componen… scores.componen… scores.componen… scores.componen…
#>               <dbl>            <dbl>            <dbl>            <dbl>
#>  1            4.58           -0.0914          0.869           -0.0176 
#>  2            3.25            0.897           0.469            0.0967 
#>  3            3.75           -0.828           0.114           -0.118  
#>  4            1.77            1.81            0.254            0.176  
#>  5            2.17            0.120          -0.00290          0.00202
#>  6            2.49           -1.78           -0.0417          -0.211  
#>  7            0.462           2.79           -0.265            0.168  
#>  8            0.543           1.01           -0.172            0.104  
#>  9            1.14           -0.673          -0.637           -0.0569 
#> 10            1.14           -2.70           -0.366           -0.185  
#> # … with 18 more rows, and 23 more variables: scores.component.5 <dbl>,
#> #   scores.component.6 <dbl>, scores.component.7 <dbl>,
#> #   scores.component.8 <dbl>, scores.component.9 <dbl>,
#> #   scores.component.10 <dbl>, scores.component.11 <dbl>,
#> #   scores.component.12 <dbl>, scores.component.13 <dbl>,
#> #   scores.component.14 <dbl>, scores.component.15 <dbl>,
#> #   scores.component.16 <dbl>, scores.component.17 <dbl>,
#> #   scores.component.18 <dbl>, scores.component.19 <dbl>,
#> #   scores.component.20 <dbl>, scores.component.21 <dbl>,
#> #   scores.component.22 <dbl>, scores.component.23 <dbl>,
#> #   scores.component.24 <dbl>, scores.component.25 <dbl>,
#> #   scores.component.26 <dbl>, scores.component.27 <dbl>

tidy_up(model, "score_cor")

#> # A tibble: 7,236 x 3
#>    xvariables components  scores.to.var.corr
#>    <chr>      <chr>                    <dbl>
#>  1 NIR1       component 1             0.246 
#>  2 NIR2       component 1             0.461 
#>  3 NIR3       component 1             0.546 
#>  4 NIR4       component 1             0.544 
#>  5 NIR5       component 1             0.480 
#>  6 NIR6       component 1             0.375 
#>  7 NIR7       component 1             0.239 
#>  8 NIR8       component 1             0.0818
#>  9 NIR9       component 1            -0.0776
#> 10 NIR10      component 1            -0.213 
#> # … with 7,226 more rows

tidy_up(model, "explained_var")

#> # A tibble: 27 x 2
#>    components   explained.variance
#>    <chr>                     <dbl>
#>  1 component 1            46.8    
#>  2 component 2            51.5    
#>  3 component 3             1.08   
#>  4 component 4             0.215  
#>  5 component 5             0.184  
#>  6 component 6             0.118  
#>  7 component 7             0.00842
#>  8 component 8             0.00655
#>  9 component 9             0.00271
#> 10 component 10            0.00135
#> # … with 17 more rows