Construct the results of the analysis

Construct the results of the analysis

construct(data, ..., na.rm = TRUE)

Arguments

data
The blueprint data object.

...
Additional args.

na.rm
Whether to remove missing values.

Value

Uses the blueprint to construct the results of the statistical analysis.

Examples

library(magrittr)
design(iris, 'cor') %>%
 add_settings() %>%
 add_variables('xvars', c('Sepal.Length', 'Sepal.Width')) %>%
 construct()

#> # Analysis for cor constructed but has not been scrubbed. 
#> # Here is a peek at the results:
#> # A tibble: 2 x 3
#>      Variables Sepal.Length Sepal.Width
#>          <chr>        <dbl>       <dbl>
#> 1 Sepal.Length    1.0000000          NA
#> 2  Sepal.Width   -0.1175698           1

design(iris, 't.test') %>%
 add_settings() %>%
 add_variables('yvars', c('Sepal.Length', 'Sepal.Width')) %>%
 add_variables('xvars', c('Petal.Length', 'Petal.Width')) %>%
 construct()

#> # Analysis for t.test constructed but has not been scrubbed. 
#> # Here is a peek at the results:
#> # A tibble: 4 x 12
#>         Yterms       Xterms   estimate estimate1 estimate2 statistic
#>          <chr>        <chr>      <dbl>     <dbl>     <dbl>     <dbl>
#> 1 Sepal.Length Petal.Length  2.0853333  5.843333  3.758000 13.098353
#> 2 Sepal.Length  Petal.Width  4.6440000  5.843333  1.199333 50.536015
#> 3  Sepal.Width Petal.Length -0.7006667  3.057333  3.758000 -4.719421
#> 4  Sepal.Width  Petal.Width  1.8580000  3.057333  1.199333 25.916013
#> # ... with 6 more variables: p.value <dbl>, parameter <dbl>, conf.low <dbl>,
#> #   conf.high <dbl>, method <fctr>, alternative <fctr>

design(iris, 'glm') %>%
 add_settings() %>%
 add_variables('yvars', c('Sepal.Length', 'Sepal.Width')) %>%
 add_variables('xvars', c('Petal.Length', 'Petal.Width')) %>%
 construct()

#> # Analysis for glm constructed but has not been scrubbed. 
#> # Here is a peek at the results:
#> # A tibble: 8 x 10
#>         Yterms       Xterms        term   estimate  std.error statistic
#>          <chr>        <chr>       <chr>      <dbl>      <dbl>     <dbl>
#> 1 Sepal.Length Petal.Length (Intercept)  4.3066034 0.07838896 54.938900
#> 2 Sepal.Length Petal.Length XtermValues  0.4089223 0.01889134 21.646019
#> 3 Sepal.Length  Petal.Width (Intercept)  4.7776294 0.07293476 65.505517
#> 4 Sepal.Length  Petal.Width XtermValues  0.8885803 0.05137355 17.296454
#> 5  Sepal.Width Petal.Length (Intercept)  3.4548745 0.07609540 45.401882
#> 6  Sepal.Width Petal.Length XtermValues -0.1057853 0.01833860 -5.768449
#> # ... with 2 more rows, and 4 more variables: p.value <dbl>, conf.low <dbl>,
#> #   conf.high <dbl>, sample.size <int>

design(iris, 'gee') %>%
 add_settings('Species') %>%
 add_variables('yvars', c('Sepal.Length', 'Sepal.Width')) %>%
 add_variables('xvars', c('Petal.Length', 'Petal.Width')) %>%
 construct()

#> Loading required namespace: geepack

#> # Analysis for gee constructed but has not been scrubbed. 
#> # Here is a peek at the results:
#> # A tibble: 8 x 12
#>         Yterms       Xterms        term   estimate  std.error    statistic
#>          <chr>        <chr>       <chr>      <dbl>      <dbl>        <dbl>
#> 1 Sepal.Length Petal.Length (Intercept)  4.3066034 0.15597956   762.314822
#> 2 Sepal.Length Petal.Length XtermValues  0.4089223 0.03261575   157.190652
#> 3 Sepal.Length  Petal.Width (Intercept)  4.7776294 0.02346810 41444.720308
#> 4 Sepal.Length  Petal.Width XtermValues  0.8885803 0.01279792  4820.749350
#> 5  Sepal.Width Petal.Length (Intercept)  3.4548745 0.19048750   328.951286
#> 6  Sepal.Width Petal.Length XtermValues -0.1057853 0.04353763     5.903657
#> # ... with 2 more rows, and 6 more variables: p.value <dbl>, conf.low <dbl>,
#> #   conf.high <dbl>, sample.total <int>, sample.max <int>, sample.min <int>

design(iris, 'pls') %>%
 add_settings() %>%
 add_variables('yvars', c('Sepal.Length', 'Sepal.Width')) %>%
 add_variables('xvars', c('Petal.Length', 'Petal.Width')) %>%
 construct()

#> Loading required namespace: pls

#> # Analysis for pls constructed but has not been scrubbed. 
#> # Here is a peek at the results:
#> Data: 	X dimension: 75 2 
#> 	Y dimension: 75 2
#> Fit method: kernelpls
#> Number of components considered: 2
#> TRAINING: % variance explained
#>               1 comps  2 comps
#> X               98.86   100.00
#> Sepal.Length    74.99    77.54
#> Sepal.Width     14.00    19.98
#> NULL