Construct the results of the analysis

Construct the results of the analysis

construct(data, ..., na.rm = TRUE)

Arguments

data
The blueprint data object.

...
Additional args.

na.rm
Whether to remove missing values.

Value

Uses the blueprint to construct the results of the statistical analysis.

Examples

library(magrittr)
design(iris, 'cor') %>%
 add_settings() %>%
 add_variables('xvars', c('Sepal.Length', 'Sepal.Width')) %>%
 construct()

#> # Analysis for cor constructed but has not been scrubbed. 
#> # Here is a peek at the results:
#> # A tibble: 2 x 3
#>   Variables    Sepal.Length Sepal.Width
#>   <chr>               <dbl>       <dbl>
#> 1 Sepal.Length        1              NA
#> 2 Sepal.Width        -0.118           1

design(iris, 't.test') %>%
 add_settings() %>%
 add_variables('yvars', c('Sepal.Length', 'Sepal.Width')) %>%
 add_variables('xvars', c('Petal.Length', 'Petal.Width')) %>%
 construct()

#> # Analysis for t.test constructed but has not been scrubbed. 
#> # Here is a peek at the results:
#> # A tibble: 4 x 12
#>   Yterms    Xterms   estimate estimate1 estimate2 statistic   p.value parameter
#>   <chr>     <chr>       <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#> 1 Sepal.Le… Petal.L…    2.09       5.84      3.76     13.1  4.26e- 29      212.
#> 2 Sepal.Le… Petal.W…    4.64       5.84      1.20     50.5  1.36e-147      296.
#> 3 Sepal.Wi… Petal.L…   -0.701      3.06      3.76     -4.72 4.98e-  6      167.
#> 4 Sepal.Wi… Petal.W…    1.86       3.06      1.20     25.9  4.11e- 71      237.
#> # ... with 4 more variables: conf.low <dbl>, conf.high <dbl>, method <chr>,
#> #   alternative <chr>

design(iris, 'glm') %>%
 add_settings() %>%
 add_variables('yvars', c('Sepal.Length', 'Sepal.Width')) %>%
 add_variables('xvars', c('Petal.Length', 'Petal.Width')) %>%
 construct()

#> # Analysis for glm constructed but has not been scrubbed. 
#> # Here is a peek at the results:
#> # A tibble: 8 x 10
#>   Yterms Xterms term  estimate std.error statistic   p.value conf.low conf.high
#>   <chr>  <chr>  <chr>    <dbl>     <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#> 1 Sepal… Petal… (Int…    4.31     0.0784     54.9  2.43e-100    4.15     4.46  
#> 2 Sepal… Petal… Xter…    0.409    0.0189     21.6  1.04e- 47    0.372    0.446 
#> 3 Sepal… Petal… (Int…    4.78     0.0729     65.5  3.34e-111    4.63     4.92  
#> 4 Sepal… Petal… Xter…    0.889    0.0514     17.3  2.33e- 37    0.788    0.989 
#> 5 Sepal… Petal… (Int…    3.45     0.0761     45.4  9.02e- 89    3.31     3.60  
#> 6 Sepal… Petal… Xter…   -0.106    0.0183     -5.77 4.51e-  8   -0.142   -0.0698
#> # ... with 2 more rows, and 1 more variable: sample.size <int>

design(iris, 'gee') %>%
 add_settings('Species') %>%
 add_variables('yvars', c('Sepal.Length', 'Sepal.Width')) %>%
 add_variables('xvars', c('Petal.Length', 'Petal.Width')) %>%
 construct()

#> Loading required namespace: geepack

#> # Analysis for gee constructed but has not been scrubbed. 
#> # Here is a peek at the results:
#> # A tibble: 8 x 12
#>   Yterms  Xterms  term  estimate std.error statistic p.value conf.low conf.high
#>   <chr>   <chr>   <chr>    <dbl>     <dbl>     <dbl>   <dbl>    <dbl>     <dbl>
#> 1 Sepal.… Petal.… (Int…    4.31     0.156     762.    0         4.00     4.61  
#> 2 Sepal.… Petal.… Xter…    0.409    0.0326    157.    0         0.345    0.473 
#> 3 Sepal.… Petal.… (Int…    4.78     0.0235  41445.    0         4.73     4.82  
#> 4 Sepal.… Petal.… Xter…    0.889    0.0128   4821.    0         0.863    0.914 
#> 5 Sepal.… Petal.… (Int…    3.45     0.190     329.    0         3.08     3.83  
#> 6 Sepal.… Petal.… Xter…   -0.106    0.0435      5.90  0.0151   -0.191   -0.0205
#> # ... with 2 more rows, and 3 more variables: sample.total <int>,
#> #   sample.max <int>, sample.min <int>