Do some final polishing of the scrubbed mason analysis data.

Do some final polishing of the scrubbed mason analysis data.

polish_renaming(data, renaming.fun, columns = NULL)

polish_filter(data, keep.pattern, column)

polish_transform_estimates(data, transform.fun)

polish_adjust_pvalue(data, method = "BH")

Arguments

data
The scrubbed object.

renaming.fun
A function, typically with base::gsub(), that searches and replaces strings.

columns
The columns to apply the renaming function to. Defaults to columns that are a factor or character vectors.

keep.pattern
Rows to keep based on a regular expression pattern.

column
The column to apply the filtering to.

transform.fun
A function to modify continuous variable columns.

method
Correction method for the p-value adjustment (stats::p.adjust()).

Functions

polish_renaming: polish_renaming simply takes a function, most likely one that uses base::gsub(), and uses that to search and replace words, etc, in the specified columns.

polish_filter: polish_filter is basically a thin wrapper around dplyr::filter(), but using base::grepl() for the pattern searching.

polish_transform_estimates: polish_transform_estimates is simply a thin wrapper around dplyr::mutate().

polish_adjust_pvalue: polish_adjust_pvalue is a thin wrapper around dplyr::mutate() and stats::p.adjust()

Examples

library(magrittr)
ds <- swiss %>%
 design('glm') %>%
 add_settings() %>%
 add_variables('yvar', c('Fertility', 'Education')) %>%
 add_variables('xvar', c('Agriculture', 'Catholic')) %>%
 add_variables('covariates', 'Examination') %>%
 construct() %>%
 scrub()
polish_renaming(ds, function(x) gsub('Education', 'Schooling', x))

#> # A tibble: 12 x 10
#>    Yterms Xterms term  estimate std.error statistic  p.value conf.low conf.high
#>    <chr>  <chr>  <chr>    <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#>  1 Ferti… Agric… (Int…  94.6       7.83       12.1  1.41e-15  79.3     110.    
#>  2 Ferti… Agric… Agri…  -0.0940    0.0859     -1.09 2.80e- 1  -0.262     0.0744
#>  3 Ferti… Agric… Exam…  -1.20      0.245      -4.89 1.40e- 5  -1.67     -0.716 
#>  4 Schoo… Agric… (Int…   7.72      5.45        1.42 1.63e- 1  -2.95     18.4   
#>  5 Schoo… Agric… Agri…  -0.128     0.0598     -2.14 3.76e- 2  -0.245    -0.0110
#>  6 Schoo… Agric… Exam…   0.591     0.170       3.47 1.16e- 3   0.258     0.925 
#>  7 Ferti… Catho… (Int…  83.0       4.98       16.7  1.21e-20  73.3      92.8   
#>  8 Ferti… Catho… Cath…   0.0418    0.0416      1.01 3.20e- 1  -0.0397    0.123 
#>  9 Ferti… Catho… Exam…  -0.886     0.217      -4.08 1.88e- 4  -1.31     -0.460 
#> 10 Schoo… Catho… (Int… -10.5       3.30       -3.20 2.56e- 3 -17.0      -4.08  
#> 11 Schoo… Catho… Cath…   0.0845    0.0275      3.07 3.70e- 3   0.0305    0.138 
#> 12 Schoo… Catho… Exam…   1.09      0.144       7.60 1.52e- 9   0.812     1.38  
#> # … with 1 more variable: sample.size <int>

polish_filter(ds, 'Xterm', 'term')

#> # A tibble: 0 x 10
#> # … with 10 variables: Yterms <chr>, Xterms <chr>, term <chr>, estimate <dbl>,
#> #   std.error <dbl>, statistic <dbl>, p.value <dbl>, conf.low <dbl>,
#> #   conf.high <dbl>, sample.size <int>

polish_adjust_pvalue(ds)[c('p.value', 'adj.p.value')]

#> # A tibble: 12 x 2
#>     p.value adj.p.value
#>       <dbl>       <dbl>
#>  1 1.41e-15    8.48e-15
#>  2 2.80e- 1    3.05e- 1
#>  3 1.40e- 5    4.20e- 5
#>  4 1.63e- 1    1.96e- 1
#>  5 3.76e- 2    5.01e- 2
#>  6 1.16e- 3    2.32e- 3
#>  7 1.21e-20    1.45e-19
#>  8 3.20e- 1    3.20e- 1
#>  9 1.88e- 4    4.52e- 4
#> 10 2.56e- 3    4.39e- 3
#> 11 3.70e- 3    5.55e- 3
#> 12 1.52e- 9    6.06e- 9

polish_transform_estimates(ds, function(x) exp(x))

#> # A tibble: 12 x 10
#>    Yterms Xterms term  estimate std.error statistic  p.value conf.low conf.high
#>    <chr>  <chr>  <chr>    <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#>  1 Ferti… Agric… (Int… 1.23e+41   2508.       12.1  1.41e-15 2.67e+34  5.64e+47
#>  2 Ferti… Agric… Agri… 9.10e- 1      1.09     -1.09 2.80e- 1 7.69e- 1  1.08e+ 0
#>  3 Ferti… Agric… Exam… 3.03e- 1      1.28     -4.89 1.40e- 5 1.87e- 1  4.89e- 1
#>  4 Educa… Agric… (Int… 2.26e+ 3    232.        1.42 1.63e- 1 5.21e- 2  9.77e+ 7
#>  5 Educa… Agric… Agri… 8.80e- 1      1.06     -2.14 3.76e- 2 7.82e- 1  9.89e- 1
#>  6 Educa… Agric… Exam… 1.81e+ 0      1.19      3.47 1.16e- 3 1.29e+ 0  2.52e+ 0
#>  7 Ferti… Catho… (Int… 1.15e+36    145.       16.7  1.21e-20 6.69e+31  1.99e+40
#>  8 Ferti… Catho… Cath… 1.04e+ 0      1.04      1.01 3.20e- 1 9.61e- 1  1.13e+ 0
#>  9 Ferti… Catho… Exam… 4.12e- 1      1.24     -4.08 1.88e- 4 2.69e- 1  6.31e- 1
#> 10 Educa… Catho… (Int… 2.63e- 5     27.0      -3.20 2.56e- 3 4.10e- 8  1.69e- 2
#> 11 Educa… Catho… Cath… 1.09e+ 0      1.03      3.07 3.70e- 3 1.03e+ 0  1.15e+ 0
#> 12 Educa… Catho… Exam… 2.99e+ 0      1.15      7.60 1.52e- 9 2.25e+ 0  3.96e+ 0
#> # … with 1 more variable: sample.size <int>