Do some final polishing of the scrubbed mason analysis data.

Do some final polishing of the scrubbed mason analysis data.

polish_renaming(data, renaming.fun, columns = NULL)

polish_filter(data, keep.pattern, column)

polish_transform_estimates(data, transform.fun)

polish_adjust_pvalue(data, method = "BH")

Arguments

data
The scrubbed object.

renaming.fun
A function, typically with gsub, that searches and replaces strings.

columns
The columns to apply the renaming function to. Defaults to columns that are a factor or character vectors.

keep.pattern
Rows to keep based on a regular expression pattern.

column
The column to apply the filtering to.

transform.fun
A function to modify continuous variable columns.

method
Correction method for the p-value adjustment (p.adjust).

Functions

polish_renaming: polish_renaming simply takes a function, most likely one that uses gsub, and uses that to search and replace words, etc, in the specified columns.

polish_filter: polish_filter is basically a thin wrapper around filter, but using grepl for the pattern searching.

polish_transform_estimates: polish_transform_estimates is simply a thin wrapper around mutate_at.

polish_adjust_pvalue: polish_adjust_pvalue is a thin wrapper around mutate and p.adjust

Examples

library(magrittr)
ds <- swiss %>%
 design('glm') %>%
 add_settings() %>%
 add_variables('yvar', c('Fertility', 'Education')) %>%
 add_variables('xvar', c('Agriculture', 'Catholic')) %>%
 add_variables('covariates', 'Examination') %>%
 construct() %>%
 scrub()
polish_renaming(ds, function(x) gsub('Education', 'Schooling', x))

#> # A tibble: 12 x 10
#>       Yterms      Xterms        term     estimate  std.error statistic
#>        <chr>       <chr>       <chr>        <dbl>      <dbl>     <dbl>
#>  1 Schooling Agriculture (Intercept)   7.72130422 5.44699896  1.417534
#>  2 Schooling Agriculture     <-Xterm  -0.12816072 0.05978132 -2.143825
#>  3 Schooling Agriculture Examination   0.59128951 0.17018382  3.474417
#>  4 Schooling    Catholic (Intercept) -10.54608842 3.29743703 -3.198268
#>  5 Schooling    Catholic     <-Xterm   0.08445744 0.02754615  3.066035
#>  6 Schooling    Catholic Examination   1.09463965 0.14399909  7.601712
#>  7 Fertility Agriculture (Intercept)  94.60968827 7.82710709 12.087440
#>  8 Fertility Agriculture     <-Xterm  -0.09399751 0.08590323 -1.094226
#>  9 Fertility Agriculture Examination  -1.19502875 0.24454695 -4.886705
#> 10 Fertility    Catholic (Intercept)  83.03565524 4.97729702 16.682881
#> 11 Fertility    Catholic     <-Xterm   0.04179341 0.04157937  1.005148
#> 12 Fertility    Catholic Examination  -0.88618608 0.21735858 -4.077070
#> # ... with 4 more variables: p.value <dbl>, conf.low <dbl>, conf.high <dbl>,
#> #   sample.size <int>

polish_filter(ds, 'Xterm', 'term')

#> # A tibble: 4 x 10
#>      Yterms      Xterms    term    estimate  std.error statistic     p.value
#>       <chr>       <chr>   <chr>       <dbl>      <dbl>     <dbl>       <dbl>
#> 1 Education Agriculture <-Xterm -0.12816072 0.05978132 -2.143825 0.037612082
#> 2 Education    Catholic <-Xterm  0.08445744 0.02754615  3.066035 0.003700765
#> 3 Fertility Agriculture <-Xterm -0.09399751 0.08590323 -1.094226 0.279810261
#> 4 Fertility    Catholic <-Xterm  0.04179341 0.04157937  1.005148 0.320321713
#> # ... with 3 more variables: conf.low <dbl>, conf.high <dbl>, sample.size <int>

polish_adjust_pvalue(ds)[c('p.value', 'adj.p.value')]

#> # A tibble: 12 x 2
#>         p.value  adj.p.value
#>           <dbl>        <dbl>
#>  1 1.633700e-01 1.960439e-01
#>  2 3.761208e-02 5.014944e-02
#>  3 1.162350e-03 2.324701e-03
#>  4 2.563100e-03 4.393885e-03
#>  5 3.700765e-03 5.551148e-03
#>  6 1.515559e-09 6.062236e-09
#>  7 1.413026e-15 8.478156e-15
#>  8 2.798103e-01 3.052476e-01
#>  9 1.400471e-05 4.201413e-05
#> 10 1.207314e-20 1.448777e-19
#> 11 3.203217e-01 3.203217e-01
#> 12 1.881458e-04 4.515498e-04

polish_transform_estimates(ds, function(x) exp(x))

#> # A tibble: 12 x 10
#>       Yterms      Xterms        term     estimate   std.error statistic
#>        <chr>       <chr>       <chr>        <dbl>       <dbl>     <dbl>
#>  1 Education Agriculture (Intercept) 2.255900e+03  232.060697  1.417534
#>  2 Education Agriculture     <-Xterm 8.797120e-01    1.061604 -2.143825
#>  3 Education Agriculture Examination 1.806316e+00    1.185523  3.474417
#>  4 Education    Catholic (Intercept) 2.629614e-05   27.043239 -3.198268
#>  5 Education    Catholic     <-Xterm 1.088127e+00    1.027929  3.066035
#>  6 Education    Catholic Examination 2.988106e+00    1.154883  7.601712
#>  7 Fertility Agriculture (Intercept) 1.225930e+41 2507.664410 12.087440
#>  8 Fertility Agriculture     <-Xterm 9.102850e-01    1.089701 -1.094226
#>  9 Fertility Agriculture Examination 3.026953e-01    1.277043 -4.886705
#> 10 Fertility    Catholic (Intercept) 1.153259e+36  145.081698 16.682881
#> 11 Fertility    Catholic     <-Xterm 1.042679e+00    1.042456  1.005148
#> 12 Fertility    Catholic Examination 4.122250e-01    1.242790 -4.077070
#> # ... with 4 more variables: p.value <dbl>, conf.low <dbl>, conf.high <dbl>,
#> #   sample.size <int>